+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Issue #275: Unify yaml configuration files
+
 ## [0.11.1] 2023-12-24
 
 ### Changed